requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
oracledb==2.4.1
Werkzeug==3.0.3
gunicorn==22.0.0
